refactor(db): move last_message_per_workspace tests from mod.rs to db/chat.rs
